requirements.txt
paho-mqtt>=1.1
wheel>=0.24.0
httplib2>=0.9.1
